Move template haskell into Utils
