+++ src/Garn/Utils.hs
@@ -1,24 +1,14 @@
+{-# LANGUAGE TemplateHaskell #-}
+
 module Garn.Utils where
 
-import Data.Aeson (FromJSON, eitherDecodeFileStrict')
 import Data.String.Conversions (cs)
 import Debug.Trace (trace)
-import GHC.Generics (Generic)
-import Language.Haskell.TH (Exp (LitE), Lit (StringL), Q, runIO)
+import Garn.ImportVersion (garnVersionSplice)
 import Text.Pretty.Simple (pShow)
 
 dbg :: (Show a) => a -> a
 dbg a = trace (cs $ pShow a) a
 
-newtype Version = Version
-  { tsLibVersion :: String
-  }
-  deriving stock (Generic)
-  deriving anyclass (FromJSON)
-
-garnVersion :: Q Exp
-garnVersion = do
-  version <- runIO $ eitherDecodeFileStrict' "./ts/internal/version.json"
-  case version of
-    Right version -> pure $ LitE $ StringL $ tsLibVersion version
-    Left err -> fail err
+garnVersion :: String
+garnVersion = $(garnVersionSplice)
